Blocking queue aggregator
