--- src/main/java/com/feedreader/rssaggregator/RSSFeedParser.java
@@ -1,136 +1,0 @@
-package com.feedreader.rssaggregator;
-
-import com.feedreader.rssaggregator.model.Feed;
-import com.feedreader.rssaggregator.model.FeedAggregate;
-import com.feedreader.rssaggregator.model.FeedMessage;
-
-import javax.xml.stream.XMLEventReader;
-import javax.xml.stream.XMLInputFactory;
-import javax.xml.stream.XMLStreamException;
-import javax.xml.stream.events.Characters;
-import javax.xml.stream.events.XMLEvent;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
-
-public class RSSFeedParser implements Runnable {
-    private static final String TITLE = "title";
-    private static final String DESCRIPTION = "description";
-    //    private static final String CHANNEL = "channel";
-    private static final String LANGUAGE = "language";
-    private static final String COPYRIGHT = "copyright";
-    private static final String LINK = "link";
-    private static final String AUTHOR = "author";
-    private static final String ITEM = "item";
-    private static final String PUB_DATE = "pubDate";
-    private static final String GUID = "guid";
-
-    private final URL url;
-    private FeedAggregate aggregate;
-
-    public RSSFeedParser(String feedUrl, FeedAggregate aggregate) {
-        this.aggregate = aggregate;
-        try {
-            this.url = new URL(feedUrl);
-        } catch (MalformedURLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public Feed readFeed() {
-        Feed feed = null;
-        try {
-            boolean isFeedHeader = true;
-            String description = "";
-            String title = "";
-            String link = "";
-            String language = "";
-            String copyright = "";
-            String author = "";
-            String pubdate = "";
-            String guid = "";
-
-            XMLInputFactory inputFactory = XMLInputFactory.newInstance();
-            InputStream in = read();
-            XMLEventReader eventReader = inputFactory.createXMLEventReader(in);
-
-            while (eventReader.hasNext()) {
-                XMLEvent event = eventReader.nextEvent();
-                if (event.isStartElement()) {
-                    String localPart = event.asStartElement().getName().getLocalPart();
-
-                    switch (localPart) {
-                        case ITEM:
-                            if (isFeedHeader) {
-                                isFeedHeader = false;
-                                feed = new Feed(title, link, description, language, copyright, pubdate);
-                            }
-                            break;
-                        case TITLE:
-                            title = getCharacterData(eventReader);
-                            break;
-                        case DESCRIPTION:
-                            description = getCharacterData(eventReader);
-                            break;
-                        case LINK:
-                            link = getCharacterData(eventReader);
-                            break;
-                        case GUID:
-                            guid = getCharacterData(eventReader);
-                            break;
-                        case LANGUAGE:
-                            language = getCharacterData(eventReader);
-                            break;
-                        case AUTHOR:
-                            author = getCharacterData(eventReader);
-                            break;
-                        case PUB_DATE:
-                            pubdate = getCharacterData(eventReader);
-                            break;
-                        case COPYRIGHT:
-                            copyright = getCharacterData(eventReader);
-                            break;
-                    }
-                } else if (event.isEndElement()) {
-                    if (event.asEndElement().getName().getLocalPart().equals(ITEM)) {
-                        FeedMessage message = new FeedMessage();
-                        message.setAuthor(author);
-                        message.setDescription(description);
-                        message.setGuid(guid);
-                        message.setLink(link);
-                        message.setTitle(title);
-//                        feed.getMessages().add(message);
-                        aggregate.addFeedMessage(message);
-                    }
-                }
-            }
-        } catch (XMLStreamException e) {
-            throw new RuntimeException(e);
-        }
-        return feed;
-    }
-
-    private String getCharacterData(XMLEventReader eventReader)
-            throws XMLStreamException {
-        String result = "";
-        XMLEvent event = eventReader.nextEvent();
-        if (event instanceof Characters) {
-            result = event.asCharacters().getData();
-        }
-        return result;
-    }
-
-    private InputStream read() {
-        try {
-            return url.openStream();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Override
-    public void run() {
-        readFeed();
-    }
-}
