+++ src/main/java/com/feedreader/rssaggregator/model/FeedMessage.java
@@ -1,12 +1,22 @@
 package com.feedreader.rssaggregator.model;
 
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class FeedMessage {
+    private static final DateFormat PUBDATE_FORMATTER = new SimpleDateFormat("EEE, dd MMM yyyy HH:mm:ss zzz");
 
     String title;
     String description;
-    String link;
     String author;
     String guid;
+    Date pubDate;
+    URL link;
+    URL source;
 
     public String getTitle() {
         return title;
@@ -25,11 +35,12 @@ public class FeedMessage {
     }
 
     public String getLink() {
-        return link;
+        return link.toString();
     }
 
-    public void setLink(String link) {
-        this.link = link;
+    public void setLink(String link) throws MalformedURLException {
+        this.link = new URL(link);
+        this.source = new URL(this.link.getProtocol() + "://" + this.link.getHost());
     }
 
     public String getAuthor() {
@@ -46,6 +57,18 @@ public class FeedMessage {
 
     public void setGuid(String guid) {
         this.guid = guid;
+    }
+
+    public Date getPubDate() {
+        return pubDate;
+    }
+
+    public void setPubDate(String pubDate) throws ParseException {
+        this.pubDate = PUBDATE_FORMATTER.parse(pubDate);
+    }
+
+    public URL getSource() {
+        return source;
     }
 
     @Override
